move status icon function to status.tsx and return JSX instead of emoji (#107)
