+++ src/frontend/containers/devPanel/panel.tsx
@@ -16,7 +16,7 @@ import useConfigContext from "@frontend/hooks/contexts/useConfigContext";
 import useDatabaseContext from "@frontend/hooks/contexts/useDatabaseContext";
 import useStatus from "@frontend/hooks/useStatus";
 import { useObservable } from "@frontend/utils";
-import { getStatusIcon } from "@frontend/utils/index";
+import { getStatusIcon } from "@frontend/utils/status";
 
 import services from "@services/index";
